So many things. Everything but persistence is done, which is cursed.
